backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
# python-jose[cryptography]==3.3.0  # JWT 관련 (POC에서는 사용하지 않음)
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
web3==6.11.3
eth-account==0.9.0
cryptography>=41.0.0
Pillow==10.1.0
aiofiles==23.2.1 